Conexión base de datos y back

+++ backend/src/main/java/com/apparquear/dao/ReservationDAO.java
@@ -4,5 +4,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.apparquear.model.Reservation;
 
-public interface ReservationDAO extends JpaRepository<Reservation, Long> {
+public interface ReservationDAO extends JpaRepository<Reservation, Integer> {
 }
